WIP: slug uniqueness migration and tests
